+++ CHANGELOG.md
@@ -1,14 +1,20 @@
 ## CHANGELOG
 
-### v0.7.2
+## [v0.7.3] - 2022-06-01
 
+### Changed
+1. Updated for compatibility with [uhppoted-lib](https://github.com/uhppoted/uhppoted-lib) v0.7.3
+
+### [v0.7.2]
+
+### Changed
 1. Migrated to uhppoted-lib `config` command implementation
 2. Reworked `get-events` to return `first`, `last` and `current` event indices.
 3. Reworked `get-event` to return `first`, `last`, `current` and `next` events.
 
-### v0.7.1
+### [v0.7.1]
 
+### Changed
 1. Task list support:
    -  `PutTaskList`
-
 2. Migrated to IUHPPOTED interface + implementation